requirements.txt
tensorflow
Flask
numpy
requests
pillow
opencv-python